pyproject.toml
[project]
name = "hackmit2025"
version = "0.1.0"
description = "Live political video fact-checker with real-time claim verification"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langgraph>=0.2.0",
    "composio-langchain>=0.6.0",
    "crawl4ai>=0.4.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "typing-extensions>=4.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]",
    "websockets",
    "redis[hiredis]",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["claim_verification"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]
